Refactor native wrapper process runner
