+++ CMakeLists.txt
@@ -510,6 +510,7 @@ set(COPPERFIN_RUNTIME_PIPELINE_SOURCES
     src/runtime/runtime_pipeline_library_export_metadata.cpp
     src/runtime/runtime_pipeline_library_export_artifacts.cpp
     src/runtime/runtime_pipeline_library_export_manifest.cpp
+    src/runtime/runtime_pipeline_native_wrapper_process.cpp
     src/runtime/runtime_pipeline_fxp_and_archive_manifest.cpp
     src/runtime/runtime_pipeline_ast_ir_manifest.cpp
     src/runtime/runtime_pipeline_csharp_and_launcher.cpp
